TA-224: Refactor tests and fix gradle hooks
