+++ Jenkinsfile_CNP
@@ -80,6 +80,6 @@ withPipeline(type, product, component) {
     }
 
     onMaster() {
-        enablePactAs([AppPipelineDsl.PactRoles.CONSUMER, AppPipelineDsl.PactRoles.PROVIDER])
+       enablePactAs([AppPipelineDsl.PactRoles.PROVIDER])
     }
 }
