+++ src/contractTest/java/uk/gov/hmcts/reform/professionalapi/BaseProviderTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.professionalapi.provider;
+package uk.gov.hmcts.reform.professionalapi;
 
 import au.com.dius.pact.provider.junit5.PactVerificationContext;
 import au.com.dius.pact.provider.junit5.PactVerificationInvocationContextProvider;
@@ -13,7 +13,7 @@ import org.springframework.test.context.junit.jupiter.SpringExtension;
 @PactBroker(scheme = "${PACT_BROKER_SCHEME:http}",
     host = "${PACT_BROKER_URL:localhost}",
     port = "${PACT_BROKER_PORT:80}", consumerVersionSelectors = {
-    @VersionSelector(tag = "${PACT_BRANCH_NAME:Dev}")})
+    @VersionSelector(tag = "master")})
 @IgnoreNoPactsToVerify
 public abstract class BaseProviderTest {
